pyproject.toml
[tool.poetry]
name = "back-garden"
version = "0.1.0"
description = "Share your garden, or theirs!"
authors = ["Victoire, Quentin, Isciane, Constant"]
license = "MIT"
readme = "README.md"
packages = [{include = "core"}]

[tool.poetry.dependencies]
python = "^3.8.1"
Django = "^4.1.2"
autopep8 = "^1.7.0"
factory-boy = "^3.2.1"
Faker = "^14.2.0"
pytest = "^7.1.3"
django-environ = "^0.9.0"
djangorestframework = "^3.14.0"
django-registration = "^3.3"
django-rest-passwordreset = "^1.3.0"
pre-commit = "^2.20.0"
psycopg2 = "^2.9.5"
django-cors-headers = "^3.13.0"
pillow = "^9.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
